+++ src/plonemeeting/restapi/config.py
@@ -5,3 +5,5 @@ CONFIG_ID_NOT_FOUND_ERROR = 'The given "config_id" named "%s" was not found!'
 IN_NAME_OF_UNAUTHORIZED = (
     'User must be Manager/MeetingManager to use "in_name_of=%s" option!'
 )
+ANNEXES_FILTER_VALUES = [
+    "to_print", "confidential", "publishable", "to_sign", "signed"]